% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdCCA_hdf5_rcpp}
\alias{bdCCA_hdf5_rcpp}
\title{CCA in hdf5 data files}
\usage{
bdCCA_hdf5_rcpp(
  filename,
  datasetX,
  datasetY,
  bcenter,
  bscale,
  mblocks,
  overwrite,
  threads = NULL
)
}
\arguments{
\item{filename, }{Character array, indicating the name of the file with data to
perform computations}

\item{datasetX, }{Character array, indicating the input datasets to be used
as X matrix}

\item{datasetY, }{Character array, indicating the input datasets to be used
as Y vector}

\item{bcenter}{logical. If TRUE (default), the data is centered by
subtracting the column means (ignoring NAs) of \code{x} from their corresponding
columns. If FALSE, no centering is performed.}

\item{mblocks}{number of equally sized blocks into which each input matrix
is partitioned for sequential processing.}

\item{overwrite, }{boolean if true, previous results in same location inside
hdf5 will be overwritten.}

\item{threads}{optional parameter. Integer with numbers of threads to be used}

\item{bscale.}{If TRUE, the data is scaled by dividing the (centered)
columns of \code{x} by their standard deviations if \code{bcenter} is TRUE, or by the
root mean square otherwise. If FALSE, no scaling is performed.}

\item{byrows}{logical if TRUE, centering is done by subtracting the rows
means, util when working with hdf5 datasets stored
in Row Major format.}
}
\description{
Compute Canonical Correlation Analyses (CCA) in data stored in hdf5 data files,
this function is a dummy versions using only c++ functions from common c++ API's
and BigDataStatMeth library
}
