\name{BDStatMethExamples-package}
\alias{BDStatMethExamples-package}
\alias{BDStatMethExamples}
\docType{package}
\title{
\packageTitle{BDStatMethExamples}
}
\description{
\packageDescription{BDStatMethExamples}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BDStatMethExamples}
\packageIndices{BDStatMethExamples}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BDStatMethExamples}

Maintainer: \packageMaintainer{BDStatMethExamples}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
