% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bdCCA.R
\name{plot_bdCCA}
\alias{plot_bdCCA}
\title{Plot Canonical Correlation Analysis}
\usage{
plot_bdCCA(
  hdf5_filename,
  metadata,
  pheno,
  column_id = NULL,
  plot_filename = NULL
)
}
\arguments{
\item{metadata}{dataset with metadata file name where dataset to normalize is stored.}

\item{pheno}{string column name with the pheno to be used in the plot}

\item{column_id}{string (optional), metadata column name to be used to merge
data (rownames in xcoef and ycoef). If null, a merge of the datasets is
performed assuming that the data is sorted.}

\item{plot_filename}{string (optional), filename to store the plot}

\item{filename}{string file name where dataset to normalize is stored.}
}
\description{
Plot Canonical Correlation Analysis with blocks results
}
\examples{
   print ("Example in vignette")
   
}
