% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCCAComponents_hdf5.R
\name{writeCCAComponents_hdf5}
\alias{writeCCAComponents_hdf5}
\title{Title}
\usage{
writeCCAComponents_hdf5(filename, ncolsX, ncolsY)
}
\arguments{
\item{filename}{string File name that contains the results}

\item{ncolsX}{integer Columns of X}

\item{ncolsY}{integer Columns of Y}
}
\value{
Writes CCA components in hdf5 data file
}
\description{
Title
}
\examples{
    hdf5_filename <- "cca_tcga_small_rcpp.hdf5"
    ncolsX = 500
    ncols_Y = 200
    writeCCAComponents_hdf5("hdf5_filename", ncolsX, ncols_Y)
}
